fix: move manifest schema into open_range package so lint CLI works outside repo root
