+++ src/open_range/manifests/__init__.py
@@ -1,0 +1,5 @@
+"""Manifest schema and loader, distributed with the open_range package."""
+
+from open_range.manifests.schema import Manifest, load_manifest
+
+__all__ = ["Manifest", "load_manifest"]
